Update README.md
The original bold, top-level text is showing up on developer.android.com in tiles in the Samples search experience. It's pretty unappealing and looks pretty bad. I've moved the warning down under the main heading and it's also in the Maintainer's notes, still. 

If we really don't want people finding this, I suggest removing the tag(s) that expose it to our site search on DAC.